Add JS Sample method
